Wait for windows to load before acting on them
Whenever a new Stage shows, it's necessary to wait for it to load completely before moving on with the test. Otherwise, the test can fail if some of the stage elements are not loaded yet by the time they are searched by the test. This can be done with the FxRobot interrupt method, that waits for the JavaFX event queue to be empty, blocking the test execution until then