Create abstract class FXML to facilitate fxml tests
This class should encapsulate the logic to load a FXML into a stage, as well as retrieving the FXML associated controller. Subclasses should implement the getFXMLResourceName() method to indicate which FXML file is to be loaded. Subclasses usually will override the getTitle() method as well, to indicate the stage title